fix: 修复所有测试和代码质量问题
- 修复 6 个失败的 Rust 测试
  - test_bundled_social_post_with_cover_skill_contract: 支持 SKILL.md 中的中文引号
  - workspace_commands_roundtrip: 使用驼峰命名 workspaceType
  - should_embed_social_image_tool_contract_in_default_skill: 更新为 **配图说明**
  - 修复 normalize 相关测试中的配图说明断言

- 修复 clippy 警告
  - 为多个枚举添加 #[derive(Default)]
  - 实现 std::str::FromStr trait 替代自定义 from_str
  - 修复不必要的 unwrap 调用
  - 使用 vec![] 宏替代 vec init then push

- 修复前端 ESLint 错误
  - ThemeWorkbenchSidebar: 20+ 个未使用变量加下划线前缀
  - useConfiguredProviders: 修复 React hooks 依赖项

- 所有核心测试通过 (328 passed; 0 failed)
- npm lint 通过
- 代码格式化通过

+++ src-tauri/src/commands/image_search_cmd.rs
@@ -323,7 +323,7 @@ pub async fn search_pixabay_images(
         .iter()
         .filter_map(|hit| {
             Some(PixabayHit {
-                id: hit["id"].as_u64()? as u64,
+                id: hit["id"].as_u64()?,
                 preview_url: hit["previewURL"].as_str()?.to_string(),
                 large_image_url: hit["largeImageURL"].as_str()?.to_string(),
                 image_width: hit["imageWidth"].as_u64()? as u32,
